Добавил задержку появления клавиатуры в формах добавления операций.

+++ Ledger/YYGLedgerDefine.h
@@ -32,5 +32,5 @@ extern int kDatabaseMajorVersion;
 extern NSString *const kDatabaseMinorVersionKey;
 extern int kDatabaseMinorVersion;
 
-extern int kKeyboardAppearanceDelay;
+extern float kKeyboardAppearanceDelay;
